--- src/main/resources/egl/client/controller/task/TestControllerTask4.fxml
@@ -1,13 +1,0 @@
-<?import javafx.scene.control.Tab?>
-<?import javafx.scene.control.TabPane?>
-<?import javafx.scene.control.TextArea?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.control.ScrollPane?>
-<ScrollPane fx:controller="egl.client.controller.task.TestControllerTask4"
-            xmlns:fx="http://javafx.com/fxml">
-   <VBox>
-      <VBox fx:id = "tasksVBox">
-      </VBox>
-   </VBox>
-</ScrollPane>
